backend/requirements.txt
# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# GPU-Accelerated ML
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning (GPU)
torch>=2.0.0
tensorflow>=2.13.0

# Model Explainability
shap>=0.42.0
lime>=0.2.0

# Hyperparameter Tuning
optuna>=3.3.0

# API Framework
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0
Pillow>=10.0.0

# Visualization (for backend reports)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
joblib>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
PyJWT>=2.8.0

# Statistical Tests
statsmodels>=0.14.0
pingouin>=0.5.3

# LLM and dataset integrations
google-generativeai>=0.8.0
kaggle>=1.6.0
